% Auto-generated: do not edit by hand
\name{''QueryBuilder}

\alias{''QueryBuilder}

\title{QueryBuilder component}

\description{
QueryBuilder is an QueryBuilder component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''QueryBuilder(id=NULL, label=NULL, value=NULL, columns=NULL,
Parameterarray=NULL, conditionArray=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Logical | numeric | character | named list | unnamed list. A label that will be printed when this component is rendered.}

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}

\item{columns}{Logical | numeric | character | named list | unnamed list. The columns displayed in the input.}

\item{Parameterarray}{Logical | numeric | character | named list | unnamed list. The Parameterarray displayed in the input.}

\item{conditionArray}{Logical | numeric | character | named list | unnamed list. The conditionArray displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

