% Auto-generated: do not edit by hand
\name{''ToggleSwitch}

\alias{''ToggleSwitch}

\title{ToggleSwitch component}

\description{

}

\usage{
''ToggleSwitch(id=NULL, checked=NULL, disabled=NULL, name=NULL,
onChange=NULL, optionLabels=NULL, small=NULL)
}

\arguments{
\item{id}{Character. }

\item{checked}{Logical. }

\item{disabled}{Logical. }

\item{name}{Character. }

\item{onChange}{}

\item{optionLabels}{Unnamed list. }

\item{small}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

